refactor, add ifs, add snow-to-liquid
