+++ pyproject.toml
@@ -13,5 +13,6 @@ dependencies = [
     "ipython>=9.1.0",
     "python-dotenv>=1.1.0",
     "rioxarray>=0.18.2",
+    "scipy>=1.15.2",
     "uvicorn>=0.34.1",
 ]
